+++ src/main/java/com/iims/placementcellservice/repository/CompanyDetailsRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompanyDetailsRepo extends JpaRepository<CompanyDetails, Long> {
+public interface CompanyDetailsRepo extends JpaRepository<CompanyDetails, Integer> {
 }
